+++ src/pages/index.ts
@@ -4,9 +4,9 @@ import ForgotPassword from "./forgot-password/forgot-password";
 import Home from "./home/home";
 import Ingredient from "./ingredient/ingredient";
 import Login from "./login/login";
+import FeedOrder from "./order/feed-order";
+import ProfileOrder from "./order/profile-order";
 import Profile from "./profile/profile";
-import FeedOrder from "./profile-order/feed-order";
-import ProfileOrder from "./profile-order/profile-order";
 import ProfileOrders from "./profile-orders/profile-orders";
 import Register from "./register/register";
 import ResetPassword from "./reset-password/reset-password";
